feat. 이번주 칼로리 합계 기능 구현 완료 (저번주 기능도 포함)

+++ domain/src/main/java/com/asetec/domain/model/state/KcalEntry.kt
@@ -1,3 +1,3 @@
 package com.asetec.domain.model.state
 
-data class KcalEntry(val date: String, val kcal: Float)
+data class KcalEntry(val date: String, val kcal: Double)
